Skybox, odrazy v oknech

+++ src/proceduralCity/Infrastructure/Structs/BuildingPartVertex.h
@@ -18,6 +18,6 @@ namespace Infrastructure
 		glm::vec3 position;
 		glm::vec3 normal;
 		glm::vec2 textureCoord;
-		short	  textureSlot;
+		int textureSlot;
 	};
 }
